refactor(checkout): cancel and success pages
